requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.6.1
pydantic[email]==2.10.3
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
openai==1.57.0
python-multipart==0.0.12
httpx==0.28.0
